+++ src/main/kotlin/me/bossm0n5t3r/leetcode/nQueens2/README.md
